Add button portfolio improvment
