+++ src/components/Portfolio/Portfolio.tsx
@@ -22,9 +22,7 @@ import { addressToTicker, initialXtoY, parseFeeToPathFee, ROUTES } from '@utils/
 import { useEffect, useMemo, useState } from 'react'
 import { useLocation, useNavigate } from 'react-router-dom'
 import { useStyles } from './style'
-
-import { SwapToken } from '@store/selectors/solanaWallet'
-import { useProcessedTokens } from '@store/hooks/userOverview/useProcessedToken'
+import { ProcessedToken } from '@store/hooks/userOverview/useProcessedToken'
 import { Overview } from './Overview/Overview/Overview'
 import { YourWallet } from './Overview/YourWallet/YourWallet'
 import { VariantType } from 'notistack'
@@ -59,7 +57,6 @@ interface IProps {
   handleClaimFee: (index: number, isLocked: boolean) => void
   handleSnackbar: (message: string, variant: VariantType) => void
   isBalanceLoading: boolean
-  tokensList: SwapToken[]
   shouldDisable: boolean
   positionListAlignment: LiquidityPools
   setPositionListAlignment: (val: LiquidityPools) => void
@@ -68,6 +65,8 @@ interface IProps {
 
   setSelectedFilters: (token: ISearchToken[]) => void
   selectedFilters: ISearchToken[]
+  isProcesing: boolean
+  processedTokens: ProcessedToken[]
 }
 
 const Portfolio: React.FC<IProps> = ({
@@ -88,12 +87,13 @@ const Portfolio: React.FC<IProps> = ({
   handleLockPosition,
   handleClosePosition,
   handleClaimFee,
-  tokensList,
   lockedLength,
   positionListAlignment,
   setPositionListAlignment,
   overviewSelectedTab,
-  handleOverviewSwitch
+  handleOverviewSwitch,
+  isProcesing,
+  processedTokens
 }) => {
   const { classes, cx } = useStyles()
 
@@ -104,11 +104,6 @@ const Portfolio: React.FC<IProps> = ({
   const isDownLg = useMediaQuery(theme.breakpoints.down('lg'))
   const isMb = useMediaQuery(theme.breakpoints.down('sm'))
   const isMd = useMediaQuery(theme.breakpoints.down('md'))
-  const { processedTokens, isProcesing } = useProcessedTokens(
-    tokensList,
-    isBalanceLoading,
-    currentNetwork
-  )
 
   const setLiquidityPoolsAlignment = (val: LiquidityPools) => {
     setAlignment(val)
